TypeError bugfix

+++ mkid_detect/mkid_detect.py
@@ -223,8 +223,8 @@ class MKIDDetect:
                 measured_wvls = self.get_photon_wavelengths(wvl, self.R_map[x, y], size=len(measured_times))
 
                 if len(measured_times) > (self.sat_rate * exp_time):
-                    measured_times = measured_times[:self.sat_rate * exp_time]
-                    measured_wvls = measured_wvls[:self.sat_rate * exp_time]
+                    measured_times = measured_times[:int(self.sat_rate * exp_time)]
+                    measured_wvls = measured_wvls[:int(self.sat_rate * exp_time)]
 
                 xs = np.full(np.shape(measured_times), x)
                 ys = np.full(np.shape(measured_times), y)
